Move Config back to tactility project
