+++ app-esp/src/main.c
@@ -1,6 +1,5 @@
 #include "board_config.h"
 #include "tactility-esp.h"
-#include "services/loader/loader.h"
 
 // Apps
 #include "hello_world/hello_world.h"
@@ -13,11 +12,6 @@ extern const AppManifest wifi_manage_app;
 
 TT_UNUSED void app_main(void) {
     static const Config config = {
-        /**
-         * Auto-select a board based on the ./sdkconfig.board.* file
-         * that you copied to ./sdkconfig before you opened this project.
-         */
-        .hardware = TT_BOARD_HARDWARE,
         .apps = {
             &hello_world_app,
             &wifi_connect_app,
@@ -29,10 +23,13 @@ TT_UNUSED void app_main(void) {
         .auto_start_app_id = NULL
     };
 
-    tt_esp_init(&config);
-    tt_init(&config.apps, CONFIG_APPS_LIMIT, &config.services, CONFIG_SERVICES_LIMIT);
+    /**
+     * Auto-select a board based on the ./sdkconfig.board.* file
+     * that you copied to ./sdkconfig before you opened this project.
+     */
+    tt_esp_init(TT_BOARD_HARDWARE);
 
-    loader_start_app("desktop", true, NULL);
+    tt_init(&config);
 
     wifi_main(NULL);
 }
