+++ tactility-esp/src/tactility-esp.h
@@ -1,23 +1,19 @@
 #pragma once
 
+#include "hardare.h"
 #include "tactility.h"
-#include "devices.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#define CONFIG_APPS_LIMIT 32
-#define CONFIG_SERVICES_LIMIT 32
-
 // Forward declarations
-typedef void* ThreadId;
 typedef void (*Bootstrap)();
 typedef TouchDriver (*CreateTouchDriver)();
 typedef DisplayDriver (*CreateDisplayDriver)();
 
 typedef struct {
-    // Optional bootstrapping method
+    // Optional bootstrapping method (e.g. to turn peripherals on)
     const Bootstrap _Nullable bootstrap;
     // Required driver for display
     const CreateDisplayDriver _Nonnull display_driver;
@@ -25,15 +21,8 @@ typedef struct {
     const CreateTouchDriver _Nullable touch_driver;
 } HardwareConfig;
 
-typedef struct {
-    const HardwareConfig* hardware;
-    // List of user applications
-    const AppManifest* const apps[CONFIG_APPS_LIMIT];
-    const ServiceManifest* const services[CONFIG_SERVICES_LIMIT];
-    const char* auto_start_app_id;
-} Config;
 
-void tt_esp_init(const Config* _Nonnull config);
+void tt_esp_init(const HardwareConfig* hardware_config);
 
 #ifdef __cplusplus
 }
