+++ tactility/src/tactility.c
@@ -2,6 +2,7 @@
 
 #include "app_manifest_registry.h"
 #include "service_registry.h"
+#include "services/loader/loader.h"
 
 #define TAG "tactility"
 
@@ -19,12 +20,9 @@ static void register_system_apps() {
     tt_app_manifest_registry_add(&system_info_app);
 }
 
-static void register_user_apps(
-    const AppManifest* const* _Nonnull apps,
-    size_t apps_count
-) {
+static void register_user_apps(const AppManifest* const apps[CONFIG_APPS_LIMIT]) {
     TT_LOG_I(TAG, "Registering user apps");
-    for (size_t i = 0; i < apps_count; i++) {
+    for (size_t i = 0; i < CONFIG_APPS_LIMIT; i++) {
         const AppManifest* manifest = apps[i];
         if (manifest != NULL) {
             tt_app_manifest_registry_add(manifest);
@@ -47,12 +45,9 @@ static void start_system_services() {
     tt_service_registry_start(loader_service.id);
 }
 
-static void register_and_start_user_services(
-    const ServiceManifest* const* services,
-    size_t servics_count
-) {
+static void register_and_start_user_services(const ServiceManifest* const services[CONFIG_SERVICES_LIMIT]) {
     TT_LOG_I(TAG, "Registering and starting user services");
-    for (size_t i = 0; i < servics_count; i++) {
+    for (size_t i = 0; i < CONFIG_SERVICES_LIMIT; i++) {
         const ServiceManifest* manifest = services[i];
         if (manifest != NULL) {
             tt_service_registry_add(manifest);
@@ -64,12 +59,7 @@ static void register_and_start_user_services(
     }
 }
 
-TT_UNUSED void tt_init(
-    const AppManifest* const* _Nonnull apps,
-    size_t apps_count,
-    const ServiceManifest* const* services,
-    size_t services_count
-) {
+TT_UNUSED void tt_init(const Config* config) {
     tt_service_registry_init();
     tt_app_manifest_registry_init();
 
@@ -78,10 +68,20 @@ TT_UNUSED void tt_init(
     register_system_services();
     register_system_apps();
     // TODO: move this after start_system_services, but desktop must subscribe to app registry events first.
-    register_user_apps(apps, apps_count);
+    register_user_apps(config->apps);
 
     // Start all services
     start_system_services();
-    register_and_start_user_services(services, services_count);
+    register_and_start_user_services(config->services);
+
+
+    TT_LOG_I(TAG, "tt_init starting desktop app");
+    loader_start_app(desktop_app.id, true, NULL);
+
+    if (config->auto_start_app_id) {
+        TT_LOG_I(TAG, "tt_init aut-starting %s", config->auto_start_app_id);
+        loader_start_app(config->auto_start_app_id, true, NULL);
+    }
+
     TT_LOG_I(TAG, "tt_init complete");
 }
